Fixing decode test

+++ Src/BlockchainSharp/Encoding/Rlp.cs
@@ -30,7 +30,7 @@
 
                 if (bytes[0] == 183 + 2)
                 {
-                    length = bytes[1] << 8 + bytes[2];
+                    length = (((int)bytes[1]) << 8) + bytes[2];
                     newbytes = new byte[length];
                     Array.Copy(bytes, 3, newbytes, 0, length);
                     return newbytes;
